pyproject.toml
[tool.poetry]
name = "product-describer"
version = "0.1.0"
description = "A Python product built with Poetry"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "product_describer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.54.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
